# perf(gatsby): Support `lte` for indexed fast filters