This implements `lte` as a fast filter, meaning it can use precomputed indexes and prevent the overhead of Sift looping over all the nodes every time a filter is invoked.

## Benchmarks show a 3x improvement at scale :D

See numbers below. Note that the speed for low counts are highly affected by the webpack build. It has a fixed cost (relative to the number of pages) but due to how things are queued and the sync-blocking nature of webpack in our build pipeline the average q/s time includes the webpack time. So lower page counts get affected more by this.

I think the benchmark shows there's still room for improvement for high scale. The perf footprint at scale is still not great. But it's 3x better now, that's gotta count for something :)

This work can be reused by most other ops, like `lt`, `gt`, and `gte`. But also for `ne` (slice the ordered array before and after the nodes with given value), and `nin` (similar approach, more chunked). Those are next.

----

Master (before this PR):
```
CI=1 NBR=1 N=50000 M=2000 yarn bench
success run queries - 1050.068s - 50002/50002 47.62/s                                                                                                                             
info Done building in 1139.285 sec                                                                                                                                                

CI=1 NBR=1 N=10000 M=2000 yarn bench
success run queries - 40.306s - 10002/10002 248.15/s                                                                                                                              
info Done building in 61.461 sec                                                                                                                                                  

CI=1 NBR=1 N=5000 M=2000 yarn bench
success run queries - 16.976s - 5002/5002 294.66/s                                                                                                                                
info Done building in 29.234 sec                                                                                                                                                  

CI=1 NBR=1 N=1000 M=2000 yarn bench
success run queries - 5.805s - 1002/1002 172.62/s                                                                                                                                 
info Done building in 10.756 sec                                                                                                                                                  

CI=1 NBR=1 N=500 M=2000 yarn bench
success run queries - 5.281s - 502/502 95.06/s                                                                                                                                    
info Done building in 9.146 sec                                                                                                                                                   
```

After this PR:
```
CI=1 NBR=1 N=50000 M=2000 yarn bench
success run queries - 332.246s - 50002/50002 150.50/s                                                                                                                             
info Done building in 421.679 sec                                                                                                                                                 

CI=1 NBR=1 N=10000 M=2000 yarn bench
success run queries - 15.818s - 10002/10002 632.31/s                                                                                                                              
info Done building in 36.964 sec                                                                                                                                                  

CI=1 NBR=1 N=5000 M=2000 yarn bench
success run queries - 9.168s - 5002/5002 545.60/s                                                                                                                                 
info Done building in 21.384 sec                                                                                                                                                  

CI=1 NBR=1 N=1000 M=2000 yarn bench
success run queries - 5.489s - 1002/1002 182.53/s                                                                                                                                 
info Done building in 10.473 sec                                                                                                                                                  

CI=1 NBR=1 N=500 M=2000 yarn bench
success run queries - 5.200s - 502/502 96.54/s                                                                                                                                    
info Done building in 9.066 sec                                                                                                                                                   
```