@pvdz 2020-04-09
Regarding index rebuild time; in earlier tests rebuilding 100k+ node indexes was a sub-second process (for one filter). I try to bench this again. It's a reasonable concern.