@pvdz 2020-04-14
Benchmarking the filter index creation. Note that `lte` is heavier than `eq` because it does the first does the same as `eq` and then a second layer of nodes ordered by value. This part should cause `lte` to perform significantly worse than just `eq`, certainly at scale.

Oddly enough, I'm not seeing this back in the stats. Possibly because the duplicate work gets cached internally in nodejs somehow.

Stats below are from the `ifsc` benchmark, adapted to use `lte`. The node count is the actual number of nodes traversed when creating indexes (not necessarily equal to the total count)

```
node count = 10001                                                                                                                                                                                                                                                                                                                                                    
ensureIndexByQuery[$eq] 1: 25.609ms                                                                                                                                                                                                                                                                                                                                   
ensureIndexByQuery[$lte] 2: 17.446ms       


node count = 50001                                                                                                                                                                                                                                                                                                                                                    
ensureIndexByQuery[$eq] 1: 105.776ms                                                                                                                                                                                                                                                                                                                                  
ensureIndexByQuery[$lte] 2: 58.098ms        

node count = 100001                                                                                                                                                                                                                                                                                                                                                   
ensureIndexByQuery[$eq] 1: 203.111ms                                                                                                                                                                                                                                                                                                                                  
ensureIndexByQuery[$lte] 2: 125.323ms      

node count = 1000001                                                                                                                                                                                                                                                                                                                                                
ensureIndexByQuery[$eq] 1: 2811.222ms                                                                                                                                                                                                                                                                                                                                 
ensureIndexByQuery[$lte] 2: 1046.264ms         
```

So at half a million or more the delay might become a bit unwieldy. This is a valid concern for incremental builds so we can look into that. However, for the most case, that's not a problem.